requirements.txt
Django==3.2.9
django-cors-headers==3.10.0
PyMySQL==1.0.2
mysqlclient==2.0.3
PyJWT==2.3.0
bcrypt==3.2.0
requests==2.26.0
freezegun==1.1.0